move _transpose_xy_dims to model.py
